+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `item_pytorch_geo_unet.json` generated using the example `unet_mlm()` function, 
+  which parses a TorchGeo UNet model with `SENTINEL2_2CLASS_NC_FTW` default weights.
+- Add `MLModelExtension.from_torch` class method to convert a PyTorch model and
+  optional TorchVision-style weights into a STAC Item with ML Model Extension metadata.
+  (Fixes [#70](https://github.com/stac-extensions/mlm/issues/70)).
 - Add torch export and packaging utilities for combining a model, transforms, and MLM schema
   compliant metadata into a single `.pt2` archive.
 - Add `cpu` device type to `mlm:accelerator` and corresponding `stac_model.runtime.AcceleratorEnum`.